+++ ts/lib/message/messageTypes.ts
@@ -14,6 +14,10 @@ export type Action = PostbackAction | UrlAction | CallAction | LocationAction | 
 
 export type ActionType = 'postback' | 'url' | 'call' | 'location' | 'share';
 
+export type AlignmentType = 'left' | 'center' | 'right';
+
+export type FieldDisplayType = 'text' | 'link';
+
 export type ChannelType = 'facebook' | 'webhook' | 'slack' | 'msteams' | 'cortana' | 'websdk' | 'androidsdk' | 'iossdk' | 'twilio' | 'test'
 
 export interface BaseAction {
@@ -66,10 +70,46 @@ export interface Keyword {
   skipAutoNumber?: boolean;
 }
 
+export interface TableHeaderColumn {
+  label: string;
+  width?: number;
+  alignment?: AlignmentType;
+}
+
+export interface TableColumn {
+  value?: any;
+  alignment?: AlignmentType;
+  displayType?: FieldDisplayType;
+  linkLabel?: string;
+}
+
+export interface FormField {
+  label: string;
+  value?: any;
+  displayType?: FieldDisplayType;
+  linkLabel?: string
+}
+
+export interface TableRow {
+  fields: TableColumn[];
+}
+
+export interface Form {
+  title?: string;
+  fields: FormField[];
+}
+
+export interface PaginationInfo {
+  totalCount: number;
+  rangeSize: number;
+  rangeStart: number;
+  status?: string
+}
+
 /**
  * define message payload types
  */
-export type MessageType = 'text' | 'card' | 'attachment' | 'location' | 'postback';
+export type MessageType = 'text' | 'card' | 'attachment'| 'table'| 'form'| 'tableForm' | 'location' | 'postback';
 export type NonRawMessagePayload = TextMessage | CardMessage | AttachmentMessage | LocationMessage | PostbackMessage ;
 export type MessagePayload = NonRawMessagePayload | RawMessage;
 export interface BaseMessagePayload {
@@ -89,13 +129,38 @@ export interface TextMessage extends BaseMessagePayload {
 
 export interface CardMessage extends BaseMessagePayload {
   type: 'card';
-  layout: 'vertical' | 'horizontal',
+  layout: 'vertical' | 'horizontal';
   cards: Card[];
 }
 export interface AttachmentMessage extends BaseMessagePayload {
   type: 'attachment';
   attachment: Attachment;
 }
+
+export interface TableMessage extends BaseMessagePayload {
+  type: 'table';
+  headings: TableHeaderColumn[];
+  rows: TableRow[];
+  paginationInfo?: PaginationInfo;
+}
+
+export interface FormMessage extends BaseMessagePayload {
+  type: 'form';
+  forms: Form[];
+  formColumns?: number;
+  paginationInfo?: PaginationInfo;
+}
+
+export interface TableFormMessage extends BaseMessagePayload {
+  type: 'tableForm';
+  headings: TableHeaderColumn[];
+  rows: TableRow[];
+  forms: Form[];
+  formColumns?: number;
+  showFormButtonLabel?: string
+  paginationInfo?: PaginationInfo;
+}
+
 export interface LocationMessage extends BaseMessagePayload {
   type: 'location';
   location: Location;